event: Change _spdk_scheduler_period_set() to use usec as argument instead of ticks
That is more inline with other places we
do things like this in SPDK. Also change
argument type to uint64_t.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/5876
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot

+++ include/spdk_internal/event.h
@@ -321,10 +321,9 @@ int _spdk_scheduler_set(char *name);
 /**
  * Change current scheduling period.
  *
- * \param period New period (ticks).
- *               Use spdk_get_ticks_hz() to translate seconds to ticks.
+ * \param period New period (microseconds).
  */
-void _spdk_scheduler_period_set(uint32_t period);
+void _spdk_scheduler_period_set(uint64_t period);
 
 /*
  * Macro used to register new reactor balancer.
